refactor(website): split summoning circle component into multiple files
